Move TimeSpanElement to its own file
